refactor: `components/Icon.ts` → `utils/icons.ts`
