+++ packages/starlight/components/ContentNotice.astro
@@ -1,5 +1,5 @@
 ---
-import type { StarlightIcon } from '../components/Icons';
+import type { StarlightIcon } from '../utils/icons';
 import Icon from '../user-components/Icon.astro';
 
 interface Props {
