+++ packages/starlight/user-components/Aside.astro
@@ -1,7 +1,7 @@
 ---
 import { AstroError } from 'astro/errors';
 import Icon from './Icon.astro';
-import { AsideDefaultIcons, type StarlightIcon } from '../components/Icons';
+import { AsideDefaultIcons, type StarlightIcon } from '../utils/icons';
 import { AsideVariants, isAsideVariant, type AsideVariant } from '../integrations/asides';
 
 interface Props {
