+++ packages/starlight/user-components/Card.astro
@@ -1,6 +1,6 @@
 ---
 import Icon from './Icon.astro';
-import type { StarlightIcon } from '../components/Icons';
+import type { StarlightIcon } from '../utils/icons';
 
 interface Props {
 	icon?: StarlightIcon;
