+++ packages/starlight/user-components/Icon.astro
@@ -1,6 +1,6 @@
 ---
 import { Icon as AstroIcon } from 'astro-icon/components';
-import { Icons, isBuiltInIcon, type StarlightIcon } from '../components/Icons';
+import { Icons, isBuiltInIcon, type StarlightIcon } from '../utils/icons';
 
 // TODO(HiDeoo) global astro-icon options?
 // TODO(HiDeoo) Can we support icons from `src/icons/` in a remark plugin?
