+++ packages/starlight/user-components/TabItem.astro
@@ -1,6 +1,6 @@
 ---
 import { TabItemTagname } from './rehype-tabs';
-import type { StarlightIcon } from '../components/Icons';
+import type { StarlightIcon } from '../utils/icons';
 
 interface Props {
 	icon?: StarlightIcon;
